charity/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==4.2.5
django-filter==24.2
django-redis==5.4.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.5.1
drf-yasg==1.21.7
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
Markdown==3.6
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.1
setuptools==69.5.1
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wheel==0.43.0
